Convert existing float encoding tests
